Communicating with server using HTTP and Observables
Data storage is being moved to the server so that now we are getting events from server not from the events array that had been hardcoded.For this purpose event.service methods have been modified to use http get method and return observables.
Also event-details.component has been modified accordingly to get events data from the server.

One more problem has been fixed.Clicking on the event in search modal was changing the url but wasn't redirecting to the corresponding page.For this purpose ngOnInit method has been modified and instead of subscribing to route snapshot now we are subscribing to data to listen to resolved data changes.

Route guard has been removed as we don't need it anymore.

Routes have been modified so that instead of canActivated we use resolver.