Fixed authentication problem
After page refresh client side was showing as if user was not authenticated while on server side actually user was authenticated.Now we added a method to track the authentication status and display the client side accordingly.We are getting data from the server and mapping it to json object.Then we check response body.If it's empty return empty object, otherwise return a json object. If the object has a property userName we set it to the current user,otherwise current user stays null.